requirements.txt
anyio==3.7.0
CacheControl==0.12.11
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Deprecated==1.2.14
exceptiongroup==1.1.1
fastapi==0.95.2
firebase-admin==6.1.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httplib2==0.22.0
httptools==0.5.0
idna==3.4
jwcrypto==1.4.2
msgpack==1.0.5
oauth2client==4.1.3
passlib==1.7.4
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
PyJWT==2.7.0
pyparsing==3.0.9
Pyrebase4==4.7.0
python-dotenv==1.0.0
python-jwt==4.0.0
PyYAML==6.0
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
